Deprecated "ensureDefaultConfig"
- [+] refactor(otpverifier): move ensureDefaultConfig to deadcode.go
